pyproject.toml
[project]
name = "giskard"
readme = "README.md"
license = { text = "Apache Software License 2.0" }
version = "3.0.0a1"
description = "The testing framework dedicated to ML models, from tabular to LLMs"
authors = [{ name = "Giskard AI", email = "hello@giskard.ai" }]

# Pypi classifiers: https://pypi.org/classifiers/
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

dependencies = [
    "giskard-core>=1.0.0a1",
    "giskard-agents>=1.0.0a1",
    "giskard-checks>=1.0.0a1",
]
requires-python = ">=3.12"

[project.urls]
"Homepage" = "https://giskard.ai"
"Source Code" = "https://github.com/Giskard-AI/giskard"
"Bug Tracker" = "https://github.com/Giskard-AI/giskard/issues"
"Documentation" = "https://docs.giskard.ai/"
"Discord" = "https://discord.gg/ABvfpbu69R"
"Linkedin" = "https://www.linkedin.com/company/giskard-ai"
"Mastodon" = "https://fosstodon.org/@Giskard"
"Twitter" = "https://twitter.com/giskard_ai"

[tool.uv.workspace]
members = [
    "libs/*",
]

[tool.uv.sources]
giskard-core = { workspace = true }
giskard-agents = { workspace = true }
giskard-checks = { workspace = true }

[tool.pytest.ini_options]
markers = [
    "functional: tests that require external API calls (deselect with '-m \"not functional\"')",
]

[dependency-groups]
dev = [
    "ipykernel==7.2.0",
    "pip-audit==2.10.0",
    "pylint-pydantic==0.4.1",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-cov>=6.1,<8",
    "pytest-timeout>=2.4.0",
]
